mcp-servers/k8s-mcp/requirements.txt
fastmcp==0.1.0
PyYAML==6.0.1
asyncio-subprocess==0.1.0